requirements.txt
twilio==8.12.0
python-dotenv==1.0.1
requests==2.31.0
flask==3.0.2 